### 2025-01-09 #169 @editorialbot
[PRE REVIEW]: DistributedWorkflows.jl - A Julia interface to a task-based workflow management system.
**Submitting author:** @FiroozehDastur (<a href="http://orcid.org/0009-0000-3334-8132">Firoozeh Dastur</a>)
**Repository:** https://github.com/FiroozehDastur/DistributedWorkflows.jl.git
**Branch with paper.md** (empty if default branch): paper/juliacon2024
**Version:** v0.2.0
**Managing EiC:** Ludovic Räss

Thanks for submitting your paper to JCON @FiroozehDastur. **Currently, there isn't a JCON editor assigned** to your paper.
