### 2025-01-13 #169 @luraess
Dear @FiroozehDastur , thanks for the reviewer suggestion! I will soon contact them such that we could start the revision process. In the meantime, could you address the remaining DOI issue listed in https://github.com/JuliaCon/proceedings-review/issues/169#issuecomment-2580573853 and potentially find a few skipped ones? Thanks for submitting to JuliaCon Proceedings!
