### 2025-01-20 1b1fa12
added missing doi's
