+++ paper/ref.bib
@@ -54,6 +54,7 @@ Citation # 7
 @misc{Introduction_to_petri_nets_book,
 author = {},
 title = {{I}ntroduction to {P}etri {N}ets},
+doi = {10.1007/978-1-4471-4276-8_10},
 howpublished = {\url{https://link.springer.com/chapter/10.1007/978-1-4471-4276-8_10}}
 }
 
@@ -86,6 +87,7 @@ Citation # 11
 @misc{Understanding_Petri_nets_book,
 author = {},
 title = {{U}nderstanding {P}etri {N}ets: {M}odeling {T}echniques, {A}nalysis {M}ethods, {C}ase {S}tudies},
+doi = {10.1007/978-3-642-33278-4},
 howpublished = {\url{https://link.springer.com/book/10.1007/978-3-642-33278-4}}
 }
 
@@ -100,8 +102,8 @@ Citation # 12
 
 Citation # 13
 @inproceedings{gpispace_paper_singular,
-author = {B\"{o}hm, J. and Fr\"{u}hbis-Kr\"{u}ger, A.},
-title = {{M}assively {P}arallel {C}omputations in {A}lgebraic {G}eometry},
+author = {B\"{o}hm, Janko and Fr\"{u}hbis-Kr\"{u}ger, Anne},
+title = {Massively Parallel Computations in Algebraic Geometry},
 year = {2021},
 isbn = {9781450383820},
 publisher = {Association for Computing Machinery},
@@ -200,7 +202,7 @@ Citation # 21
    author = {Gruenewald, D. and Ettrich, N. and Rahn, M. and Pfreundt, F.J.},
    title = {{FRTM} - {A} {P}roductive {F}ramework for {R}everse {T}ime {M}igration}, 
    year = {2014},
-   doi = {https://doi.org/10.3997/2214-4609.20141914},
+   doi = {10.3997/2214-4609.20141914},
    url = {https://www.earthdoc.org/content/papers/10.3997/2214-4609.20141914},
    publisher = {European Association of Geoscientists &amp; Engineers},
    issn = {2214-4609},
@@ -300,6 +302,7 @@ Citation # 29
   title={{M}ap{R}educe in {GPI}-{S}pace},
   author={Rotaru, T. and Rahn, M. and Pfreundt, F.J.},
   booktitle={{E}uro-{P}ar 2013: {P}arallel {P}rocessing {W}orkshops: {B}igData{C}loud, {DIHC}, {F}ed{ICI}, {H}etero{P}ar, {H}i{BB}, {LSDVE}, {MHPC}, {OMHI}, {PADABS}, {PROPER}, {R}esilience, {ROME}, and {UCHPC} 2013, {A}achen, {G}ermany, {A}ugust 26-27, 2013. {R}evised {S}elected {P}apers 19},
+  doi = {10.1007/978-3-642-54420-0_5},
   pages={43--52},
   year={2014},
   organization={Springer}
