### 2025-01-20 #169 @FiroozehDastur
Dear @luraess, I fixed the missing DOI issue and added DOI's to as many of the references as I could find. The references that don't have a DOI are mainly because either they are git repositories, other Julia packages, or links to other tools that I am citing in the paper:)
