### 2025-02-14 #169 @fingolfin
Hi there! I am in principle willing, but what should a review entail? Are there guidelines somewhere?

I would also like to disclose that I am a professor at the same university as @FiroozehDastur and she was/is my teaching assistant in a Master's course I taught this term. However I had zero involvement with the package under review so far, except that when @FiroozehDastur approach me and asked me whether I thought it was a good idea to submit it to JuliaCon, I fully encouraged her to do so.

While I personally don't think this impedes my fair judgement of her, I recognize others may feel different and thus wanted to be upfront about this connection.
