### 2025-02-14 #169 @fieker
I scond the question as to what would be involved: reviewing of the paper and/or the code? Furthermore, Firoozeh is my PhD student - she works on unrelated topics, but under my supervision. I have had no involvment into the work for the paper however.
