### 2025-02-14 #169 @luraess
Thank you @fieker and @fingolfin for your prompt feedback and for providing further clarification about potential conflict of interest. I could understand that in some cases the current configuration may not the most appropriate, however, I believe that in the context of this revision it should be all fine, from my point of view (acknowledging the fact that both potential reviewers are also OK with this).

If we go on, the revision entails both code and manuscript checking. Succinct guidelines can be found here https://juliacon.github.io/proceedings-guide/reviewer/. Upon revision start, reviewers are invited to generate a checklist which would help and guide the revision. Issues relating to the online code should be raised in the form of GitHub issues directly on the software repo, while suggestions and comments regarding the manuscript could be made in the revision issue conversation.

If this sounds like the way forward for both of you, please 👍 and I will start the revision process. Thanks!
