### 2025-02-20 #181 @editorialbot
[REVIEW]: DistributedWorkflows.jl - A Julia interface to a task-based workflow management system.
**Submitting author:** @FiroozehDastur (<a href="http://orcid.org/0009-0000-3334-8132">Firoozeh Dastur</a>)
**Repository:** https://github.com/FiroozehDastur/DistributedWorkflows.jl.git
**Branch with paper.md** (empty if default branch): paper/juliacon2024
**Version:** v0.2.0
