### 2025-02-20 #181 @luraess
Hi all, thanks for following up here. Note that you can ignore the `Failed to discover a Statement of need section in paper` feedback as we do not explicitly enforce this in JCon submissions.
