### 2025-03-04 FiroozehDastur/DistributedWorkflows.jl#21 @fingolfin
Licensing concerns
The README instructs the user to run an opaque self-extracting binary.

But where do these come from? Is there documentation somewhere explaining how these `distributedworkflow_XYZ` archives are created?

I downloaded one and extracted it in a VM, and found these files:
```
$ ls -R .
.:
build_cache  install.sh  repo

./build_cache:
_pgp                                                                                                                   linux-debian11-x86_64_v3-gcc-10.2.1-hwloc-2.9.1-5j72rx2brd76nwuh66l4h626hp33ua2m.spec.json.sig
index.json                                                                                                             linux-debian11-x86_64_v3-gcc-10.2.1-libiconv-1.17-bhtfz6wxygkwewu7a5cbva3hphn5sasd.spec.json.sig
index.json.hash                                                                                                        linux-debian11-x86_64_v3-gcc-10.2.1-libpciaccess-0.17-ra663wbjkcuzkly7o7glhad3x4amcbn4.spec.json.sig
linux-debian11-x86_64_v3                                                                                               linux-debian11-x86_64_v3-gcc-10.2.1-libsigsegv-2.14-2xfhfjuaxgfg52njswafporqecwc5e7f.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-berkeley-db-18.1.40-jxqezua3vnzviwdde22dnnn4ftguo5kt.spec.json.sig                 linux-debian11-x86_64_v3-gcc-10.2.1-libssh2-1.11.0-yaq6foqrxvuyhdnpkyz2cqog5ejulx6j.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-boost-1.63.0-eelwyai253cp3ncrgfjsknky3sbtfaqi.spec.json.sig                        linux-debian11-x86_64_v3-gcc-10.2.1-libtool-2.4.7-qkcficq53dqlhoucbupl7llapl3ytdtw.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-bzip2-1.0.8-zfuxwyvajbnk5harbclgxu3kkhpsj4vu.spec.json.sig                         linux-debian11-x86_64_v3-gcc-10.2.1-m4-1.4.19-uywihodgut7dbu6du3gkejqtdjl4esmm.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-ca-certificates-mozilla-2023-05-30-ydyz5rtaeezdd7rqol6ahvzubeaydr2y.spec.json.sig  linux-debian11-x86_64_v3-gcc-10.2.1-ncurses-6.4-6yyq7ujubwm3kyn33cqfdbzwc7altrt4.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-chrpath-0.16-4vztv6mkxr6phjlp5poy4jkqbp6iqilu.spec.json.sig                        linux-debian11-x86_64_v3-gcc-10.2.1-nghttp2-1.57.0-b2lg2ibbfzu6gi3u6klnmnfgaoqhbkxc.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-cmake-3.27.7-z4w3zw2jo6mgpenvvbj7pptunn7ihzsm.spec.json.sig                        linux-debian11-x86_64_v3-gcc-10.2.1-openssl-3.1.3-mprhptd52rqhkszenfgcsosou735mklj.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-curl-8.4.0-d65tmfpfhfn547bktaksgvxgguzsggxo.spec.json.sig                          linux-debian11-x86_64_v3-gcc-10.2.1-perl-5.38.0-hhrdw2sgs4u63ig7ljid3tvxajwflayb.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-diffutils-3.9-5xnv67gkj3nz4cal66xqbh65e7rdadhu.spec.json.sig                       linux-debian11-x86_64_v3-gcc-10.2.1-pkgconf-1.9.5-lgtucd6ctfihjpvpinkfhc4gnjkg5r3p.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-distributedworkflow-latest-loymavzqg7jkyzyu46qul435e6skmpdi.spec.json.sig          linux-debian11-x86_64_v3-gcc-10.2.1-readline-8.2-xx6bjh7tgiwx4weeqaopp4hby6g6vjaw.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-executor-latest-g4fbel2s6lns2i6ro4i3qjg7sboz5kye.spec.json.sig                     linux-debian11-x86_64_v3-gcc-10.2.1-util-macros-1.19.3-jl67jcaenejiqtajjzkzz2lqqk6b6pmb.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-gdbm-1.23-bqerekzvawxbxerm3vtbpmy6w3jttecg.spec.json.sig                           linux-debian11-x86_64_v3-gcc-10.2.1-xz-5.4.1-hxiintmu5noxrkqvz5stdvmazatxm7mr.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-gmake-4.4.1-gq64uicu5a4qzdcumnxgvk3k3o2kyxch.spec.json.sig                         linux-debian11-x86_64_v3-gcc-10.2.1-zlib-ng-2.1.4-vfknuvdtychagyxjj6mt634kkadajlmc.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-gpi-space-23.06-wupua3kzmcyjvso77jqfaswlbi733i6e.spec.json.sig                     linux-debian11-x86_64_v3-gcc-10.2.1-zstd-1.5.5-xoh7ab7tjny3motqbj67qnzargapj6pm.spec.json.sig
linux-debian11-x86_64_v3-gcc-10.2.1-gspcdriver-latest-omqrjpvqwubbqln7nshb5drj54fvab2h.spec.json.sig

./build_cache/_pgp:
346575B1E8570D3A8A13DB3F8195390F0870FE68.pub  index.json

./build_cache/linux-debian11-x86_64_v3:
gcc-10.2.1

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1:
berkeley-db-18.1.40  ca-certificates-mozilla-2023-05-30  curl-8.4.0                  executor-latest  gpi-space-23.06    libiconv-1.17      libssh2-1.11.0  ncurses-6.4     perl-5.38.0    util-macros-1.19.3  zstd-1.5.5
boost-1.63.0         chrpath-0.16                        diffutils-3.9               gdbm-1.23        gspcdriver-latest  libpciaccess-0.17  libtool-2.4.7   nghttp2-1.57.0  pkgconf-1.9.5  xz-5.4.1
bzip2-1.0.8          cmake-3.27.7                        distributedworkflow-latest  gmake-4.4.1      hwloc-2.9.1        libsigsegv-2.14    m4-1.4.19       openssl-3.1.3   readline-8.2   zlib-ng-2.1.4

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/berkeley-db-18.1.40:
linux-debian11-x86_64_v3-gcc-10.2.1-berkeley-db-18.1.40-jxqezua3vnzviwdde22dnnn4ftguo5kt.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/boost-1.63.0:
linux-debian11-x86_64_v3-gcc-10.2.1-boost-1.63.0-eelwyai253cp3ncrgfjsknky3sbtfaqi.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/bzip2-1.0.8:
linux-debian11-x86_64_v3-gcc-10.2.1-bzip2-1.0.8-zfuxwyvajbnk5harbclgxu3kkhpsj4vu.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/ca-certificates-mozilla-2023-05-30:
linux-debian11-x86_64_v3-gcc-10.2.1-ca-certificates-mozilla-2023-05-30-ydyz5rtaeezdd7rqol6ahvzubeaydr2y.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/chrpath-0.16:
linux-debian11-x86_64_v3-gcc-10.2.1-chrpath-0.16-4vztv6mkxr6phjlp5poy4jkqbp6iqilu.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/cmake-3.27.7:
linux-debian11-x86_64_v3-gcc-10.2.1-cmake-3.27.7-z4w3zw2jo6mgpenvvbj7pptunn7ihzsm.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/curl-8.4.0:
linux-debian11-x86_64_v3-gcc-10.2.1-curl-8.4.0-d65tmfpfhfn547bktaksgvxgguzsggxo.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/diffutils-3.9:
linux-debian11-x86_64_v3-gcc-10.2.1-diffutils-3.9-5xnv67gkj3nz4cal66xqbh65e7rdadhu.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/distributedworkflow-latest:
linux-debian11-x86_64_v3-gcc-10.2.1-distributedworkflow-latest-loymavzqg7jkyzyu46qul435e6skmpdi.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/executor-latest:
linux-debian11-x86_64_v3-gcc-10.2.1-executor-latest-g4fbel2s6lns2i6ro4i3qjg7sboz5kye.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/gdbm-1.23:
linux-debian11-x86_64_v3-gcc-10.2.1-gdbm-1.23-bqerekzvawxbxerm3vtbpmy6w3jttecg.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/gmake-4.4.1:
linux-debian11-x86_64_v3-gcc-10.2.1-gmake-4.4.1-gq64uicu5a4qzdcumnxgvk3k3o2kyxch.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/gpi-space-23.06:
linux-debian11-x86_64_v3-gcc-10.2.1-gpi-space-23.06-wupua3kzmcyjvso77jqfaswlbi733i6e.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/gspcdriver-latest:
linux-debian11-x86_64_v3-gcc-10.2.1-gspcdriver-latest-omqrjpvqwubbqln7nshb5drj54fvab2h.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/hwloc-2.9.1:
linux-debian11-x86_64_v3-gcc-10.2.1-hwloc-2.9.1-5j72rx2brd76nwuh66l4h626hp33ua2m.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/libiconv-1.17:
linux-debian11-x86_64_v3-gcc-10.2.1-libiconv-1.17-bhtfz6wxygkwewu7a5cbva3hphn5sasd.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/libpciaccess-0.17:
linux-debian11-x86_64_v3-gcc-10.2.1-libpciaccess-0.17-ra663wbjkcuzkly7o7glhad3x4amcbn4.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/libsigsegv-2.14:
linux-debian11-x86_64_v3-gcc-10.2.1-libsigsegv-2.14-2xfhfjuaxgfg52njswafporqecwc5e7f.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/libssh2-1.11.0:
linux-debian11-x86_64_v3-gcc-10.2.1-libssh2-1.11.0-yaq6foqrxvuyhdnpkyz2cqog5ejulx6j.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/libtool-2.4.7:
linux-debian11-x86_64_v3-gcc-10.2.1-libtool-2.4.7-qkcficq53dqlhoucbupl7llapl3ytdtw.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/m4-1.4.19:
linux-debian11-x86_64_v3-gcc-10.2.1-m4-1.4.19-uywihodgut7dbu6du3gkejqtdjl4esmm.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/ncurses-6.4:
linux-debian11-x86_64_v3-gcc-10.2.1-ncurses-6.4-6yyq7ujubwm3kyn33cqfdbzwc7altrt4.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/nghttp2-1.57.0:
linux-debian11-x86_64_v3-gcc-10.2.1-nghttp2-1.57.0-b2lg2ibbfzu6gi3u6klnmnfgaoqhbkxc.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/openssl-3.1.3:
linux-debian11-x86_64_v3-gcc-10.2.1-openssl-3.1.3-mprhptd52rqhkszenfgcsosou735mklj.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/perl-5.38.0:
linux-debian11-x86_64_v3-gcc-10.2.1-perl-5.38.0-hhrdw2sgs4u63ig7ljid3tvxajwflayb.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/pkgconf-1.9.5:
linux-debian11-x86_64_v3-gcc-10.2.1-pkgconf-1.9.5-lgtucd6ctfihjpvpinkfhc4gnjkg5r3p.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/readline-8.2:
linux-debian11-x86_64_v3-gcc-10.2.1-readline-8.2-xx6bjh7tgiwx4weeqaopp4hby6g6vjaw.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/util-macros-1.19.3:
linux-debian11-x86_64_v3-gcc-10.2.1-util-macros-1.19.3-jl67jcaenejiqtajjzkzz2lqqk6b6pmb.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/xz-5.4.1:
linux-debian11-x86_64_v3-gcc-10.2.1-xz-5.4.1-hxiintmu5noxrkqvz5stdvmazatxm7mr.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/zlib-ng-2.1.4:
linux-debian11-x86_64_v3-gcc-10.2.1-zlib-ng-2.1.4-vfknuvdtychagyxjj6mt634kkadajlmc.spack

./build_cache/linux-debian11-x86_64_v3/gcc-10.2.1/zstd-1.5.5:
linux-debian11-x86_64_v3-gcc-10.2.1-zstd-1.5.5-xoh7ab7tjny3motqbj67qnzargapj6pm.spack

./repo:
packages  repo.yaml

./repo/packages:
distributedworkflow  executor  gspcdriver

./repo/packages/distributedworkflow:
package.py

./repo/packages/executor:
package.py

./repo/packages/gspcdriver:
package.py
```

Many of these files (esp. the `.spack` files) contain binary data but no clear licensing terms. But based on the name I'd guess they contain binaries for the named software, e.g. "readline-82", which is GPL, while the license for `DistributedWorkflows.jl` suggests MIT is the license.

Perhaps that's OK, perhaps not, but I think at the very least the licensing status ought to be documented clearly.

(The other issues is of course reproducibility: as long as it is not clear where the data comes from, this is also not a given)
