### 2025-03-09 FiroozehDastur/DistributedWorkflow.jl#21 @mzeyen1985
The binary distribution is a binary Spack build cache (https://spack.readthedocs.io/en/latest/binary_caches.html). The build cache was generated for GPI-Space (GPL-3) using its public Spack recipe (https://packages.spack.io/package.html?name=gpi-space), as well as for our own C++ packages distributedworkflow,  executor, and  gspcdriver. Additionally, the provided self-extracting binaries are streamlining the otherwise error-prone and cumbersome installation process of the underlying system. To ensure the protection of our intellectual property, our C++ packages are not available as open source at this time.

Using the BSD-3-Clause license for DistributedWorkflows.jl is not an issue here as the C++ part linking to GPI-Space is dynamically loaded at runtime through CxxWrap.jl.
