### 2025-03-10 FiroozehDastur/DistributedWorkflow.jl#21 @fingolfin
OK, all fair enough, but I believe this should be addressed explicitly in the README: right now a cursory look at the package may make one believe it is a pure BSD licensed project. But in fact installing it pulls in GPL-3 as well as proprietary software with unspecified licensing terms. For many companies that will be a no-go, or at least require further vetting.

Regarding your proprietary code: so, under which license (not open source, obviously) may one use it then?
