### 2025-07-10 5714b3e
README.md: clarify how to install binary component
The downloaded file won't have the excecutable bit set.
