+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"Serialization/custom_serializer/#Instructions-on-how-to-use-a-serializer-different-than-Julia's-inbuilt-serializer","page":"Serialization","title":"Instructions on how to use a serializer different than Julia's inbuilt serializer","text":"","category":"section"},{"location":"api/#Available-features-in-the-public-API","page":"API","title":"Available features in the public API","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"CurrentModule = DistributedWorkflows","category":"page"},{"location":"api/","page":"API","title":"API","text":"Modules = [DistributedWorkflows]\nPages = [\"wrapper.jl\"]","category":"page"},{"location":"api/#DistributedWorkflows.application_config-Tuple{String, String, String}","page":"API","title":"DistributedWorkflows.application_config","text":"application_config(port::String, impl::String, fname::String)\n\nConvenience constructor for configuring a workflow application with a single transition.\n\nArguments\n\nports::String: Ports to configure for the workflow transition.\nimpl::String: Julia file containing the implementation called by the workflow transition.\nfnames::String: Function name to be executed by the workflow transition.\n\nSee also Workflow_PetriNet, generate_workflow, compile_workflow.\n\n\n\n\n\n","category":"method"},{"location":"api/#DistributedWorkflows.application_config-Tuple{Vector{String}, String, Vector{String}}","page":"API","title":"DistributedWorkflows.application_config","text":"application_config(ports::Vector{String}, impl::String, fnames::Vector{String})\n\nConvenience constructor for configuring a workflow application with multiple transitions sourcing their implementation details from the same file.\n\nArguments\n\nports::Vector{String}: List of ports to configure for the workflow transitions.\nimpl::String: Julia file containing the implementations called by the workflow transitions.\nfnames::Vector{String}: List of function names to be executed by the workflow transitions.\n\nSee also Workflow_PetriNet, generate_workflow, compile_workflow.\n\n\n\n\n\n","category":"method"},{"location":"api/#DistributedWorkflows.application_config-Tuple{Vector{String}, Vector{String}, Vector{String}}","page":"API","title":"DistributedWorkflows.application_config","text":"application_config(ports::Vector{String}, impl::Vector{String}, fnames::Vector{String})\n\nConstructor for configuring a workflow application with multiple transitions.\n\nArguments\n\nports::Vector{String}: List of ports to configure for the workflow transitions.\nimpl::Vector{String}: List of julia files containing the implementations called by the workflow transitions.\nfnames::Vector{String}: List of function names to be executed by the workflow transitions.\n\nSee also Workflow_PetriNet, generate_workflow, compile_workflow.\n\n\n\n\n\n","category":"method"},{"location":"api/#DistributedWorkflows.client-Tuple{Int64, String, String, String, Int64}","page":"API","title":"DistributedWorkflows.client","text":"client(workers::Int, nodefile::String, rif_strategy::String, log_host::String, log_port::Int)\n\nConfigures and starts a client setting up the workflow execution infrastructure and connects to a logging service. The nodefile will be automatically populated with the local host name if it doesn't exist in the given location or the rif_st
[…]
