### 2025-09-05 d3c72d6
WIP: revisions to the README as requested by one of the reviewers
