### 2025-09-12 31c685b
WIP: paper revisions
