+++ docs/src/troubleshooting.md
@@ -1,1 +1,26 @@
 # Troubleshooting
+
+If you encounter issues after following the README instructions, try the steps below:  
+
+## Troubleshooting the Client
+
+- Ensure **Spack** is correctly installed and sourced in your environment.  
+- Verify that you have downloaded and installed the correct binary for your operating system.  
+- Confirm that the binary is properly installed and accessible.  
+- If the problem persists, consider opening a GitHub issue with:  
+  - A minimal example that reproduces the error.  
+  - Details about your operating system, environment, and steps taken prior to the issue.  
+
+## Troubleshooting DistributedWorkflows
+
+- Make sure you are using a supported version of Julia.  
+- Test one of the provided minimal examples to verify that they run correctly.  
+- If the problem continues, you may open a GitHub issue including:  
+  - A minimal reproducible example.  
+  - Information about your environment and the steps leading to the error.  
+
+## Other Issues
+
+- For any other problems, you can open a GitHub issue with:  
+  - A minimal reproducible example.  
+  - Relevant details about your setup and the steps leading to the error.  
