### 2025-10-17 75aa2e2
updates to the documentation to fulfill paper requirements
