+++ docs/src/Serialization/custom_serializer.md
@@ -1,1 +1,3 @@
-# Instructions on how to use a serializer different than Julia's inbuilt serializer
+# Serialization
+
+For examples on how to use a serializer different than Julia's inbuilt serializer see [here](https://github.com/FiroozehDastur/DistributedWorkflows.jl/tree/main/examples/other_serializers).
