+++ docs/src/local_launch.md
@@ -1,1 +1,0 @@
-# Locally test your applications
