+++ docs/src/PetriNet/PetriNet.md
@@ -1,10 +1,0 @@
-# Workflow
-
-```@meta
-CurrentModule = DistributedWorkflows
-```
-
-```@autodocs
-Modules = [DistributedWorkflows]
-Pages = ["petri_net.jl", "workflow_compiler.jl", "workflow_renderer.jl", "xml_generator.jl"]
-```
