### 2025-10-20 26634c0
disclaimer added for Julia version compatibility
