+++ v0.2.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"Serialization/custom_serializer/#Serialization","page":"Serialization","title":"Serialization","text":"","category":"section"},{"location":"Serialization/custom_serializer/","page":"Serialization","title":"Serialization","text":"For examples on how to use a serializer different than Julia's inbuilt serializer see here.","category":"page"},{"location":"api/#Available-Features","page":"Available Features","title":"Available Features","text":"","category":"section"},{"location":"api/","page":"Available Features","title":"Available Features","text":"This documentation is divided into two main sections:","category":"page"},{"location":"api/","page":"Available Features","title":"Available Features","text":"Public API — methods from GPI-Space made accessible in Julia.\nWorkflow — tools and utilities for defining, compiling, and visualizing workflows.","category":"page"},{"location":"api/","page":"Available Features","title":"Available Features","text":"","category":"page"},{"location":"api/#Public-API","page":"Available Features","title":"Public API","text":"","category":"section"},{"location":"api/","page":"Available Features","title":"Available Features","text":"This section presents a collection of methods originating from GPI-Space (the underlying workflow management system) and made accessible in Julia through simple wrappers.","category":"page"},{"location":"api/","page":"Available Features","title":"Available Features","text":"For functionality related to workflow generation, see the Workflow section below.","category":"page"},{"location":"api/#DistributedWorkflows.application_config-Tuple{String, String, String}","page":"Available Features","title":"DistributedWorkflows.application_config","text":"application_config(port::String, impl::String, fname::String)\n\nConvenience constructor for configuring a workflow application with a single transition.\n\nArguments\n\nports::String: Ports to configure for the workflow transition.\nimpl::String: Julia file containing the implementation called by the workflow transition.\nfnames::String: Function name to be executed by the workflow transition.\n\nSee also Workflow_PetriNet, generate_workflow, compile_workflow.\n\n\n\n\n\n","category":"method"},{"location":"api/#DistributedWorkflows.application_config-Tuple{Vector{String}, String, Vector{String}}","page":"Available Features","title":"DistributedWorkflows.application_config","text":"application_config(ports::Vector{String}, impl::String, fnames::Vector{String})\n\nConvenience constructor for configuring a workflow application with multiple transitions sourcing their implementation details from the same file.\n\nArguments\n\nports::Vector{String}: List of ports to configure for the workflow transitions.\nimpl::String: Julia file containing the implementations called by the workflow transitions.\nfnames::Vector{String}: List of function names to be executed by the workflow transitions.\n\nSee also Workflow_PetriNet, generate_workflow, compile_workflow.\n\n\n\n\n\n","category":"method"},{"location":"api/#DistributedWorkflows.application_config-Tuple{Vector{String}, Vector{String}, Vector{String}}","page":"Available Features","title":"DistributedWorkflows.application_config","text":"application_config(ports::Vector{String}, impl::Vector{String}, fnames::Vector{String})\n\nConstructor for configuring a workflow application with multiple transitions.\n\nArguments\n\nports::Vector{String}
[…]
