+++ README.md
@@ -5,14 +5,14 @@
 
 [![CI](https://github.com/FiroozehDastur/DistributedWorkflows.jl/actions/workflows/CI.yml/badge.svg)](https://github.com/FiroozehDastur/DistributedWorkflows.jl/actions/workflows/CI.yml)
 [![codecov](https://codecov.io/gh/FiroozehDastur/DistributedWorkflows.jl/graph/badge.svg?token=9JIYL7YJYK)](https://codecov.io/gh/FiroozehDastur/DistributedWorkflows.jl)
-[![](https://img.shields.io/badge/docs-dev-blue.svg)](https://FiroozehDastur.github.io/DistributedWorkflows.jl/dev)
+[![](https://img.shields.io/badge/docs-stable-blue.svg)](https://firoozehdastur.github.io/DistributedWorkflows.jl/stable/)
 
 > ---
 > **⚠️ Compatibility Notice**
 >
 > This package currently supports Julia v1.9.2 or earlier, as required for the JuliaCon proceedings. A new version compatible with later Julia releases will be provided after the publication is finalized.
 >
-> For compatibility with the operating systems check the [release artifacts](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/tag/v0.1.0).
+> For compatibility with the operating systems check the [release artifacts](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/tag/v0.2.1).
 >
 > ---
 
@@ -42,7 +42,7 @@ Please cite this package as follows if you use it in your work:
    >
    > ---
 
-2. Next, download the respective binary for your system from the following: [Binaries for DistributedWorkflows.jl](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/tag/v0.1.0)
+2. Next, download the respective binary for your system from the following: [Binaries for DistributedWorkflows.jl](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/tag/v0.2.1)
 
    *OR* 
 
@@ -55,11 +55,11 @@ Please cite this package as follows if you use it in your work:
    >
    > ---
 
-   * [Ubuntu 22.04](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/download/v0.1.0/distributedworkflow_ubuntu22)
-   * [Debian 11](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/download/v0.1.0/distributedworkflow_debian11)
-   * [Debian 12](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/download/v0.1.0/distributedworkflow_debian12)
-   * [Rocky Linux 8](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/download/v0.1.0/distributedworkflow_rockylinux8)
-   * [Rocky Linux 9](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/download/v0.1.0/distributedworkflow_rockylinux9)
+   * [Ubuntu 22.04](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/download/v0.2.1/distributedworkflow_ubuntu22)
+   * [Debian 11](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/download/v0.2.1/distributedworkflow_debian11)
+   * [Debian 12](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/download/v0.2.1/distributedworkflow_debian12)
+   * [Rocky Linux 8](https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/download/v0.2.1/distributedworkflow_rockylinux8)
[…]
