### 2025-11-11 FiroozehDastur/DistributedWorkflow.jl#21 @FiroozehDastur
Thank you very much for the thoughtful feedback. You make a fair and valuable point.
The licenses for the underlying dependencies have now been included with the binaries and are mentioned in the new release notes. We’ve also added a cautionary “Compatibility Notice” in the README clarifying that the package currently supports only a specific Julia version (or lower).

Just to reiterate, this package is not intended for production or industry use. It is primarily an academic proof of concept. The accompanying paper focuses on the interface rather than the underlying dependencies. The proprietary components form part of my own intellectual property, and due to previous experiences with IP misuse and theft, I’ve decided not to make those sources publicly available.

Please let us know if these updates address your concerns. We’ll be happy to close the issue once you confirm everything looks good.
