### 2025-11-12 #181 @FiroozehDastur
Hi @luraess, apologies for the delay in getting back to you. I had some unforeseen personal matters that required my attention.

I’ve now completed all the issues raised by @fieker and @fingolfin related to the code base (awaiting their approval on the corresponding issues) and am currently addressing the remaining minor points in the paper draft.

I’ll post an update here once the revised version of the paper is uploaded. Thank you for your patience
