### 2025-11-12 5df65d1
paper revision based on the issues raised by reviewers
