+++ paper/paper.tex
@@ -35,7 +35,7 @@ Additionally, increasing data sizes and problem complexity have raised the neces
 For example, in the modeling and analysis of protein synthesis and DNA mutation, distributed systems enable efficient simulations of complex biological processes, providing insights into genetics and molecular biology (e.g. \cite{biology_example_petrinets_openaccess}).
 Similarly, weather simulations have greatly benefited from distributed systems, allowing for more accurate and faster predictions (e.g. \cite{WeatherForcast_HPC_paper}).
 Additionally, climate simulations are enabled to run over larger time ranges (e.g. \cite{TheJuly2019EuropeanHeatWaveinaWarmerClimateStorylineScenarioswithaCoupledModelUsingSpectralNudging}).
-Other examples include asteroid impact simulations\cite{asteroid_vis_paper}, black hole mergers\cite{blackhole_merger}, and nuclear fusion simulations\cite{in_situ_visualization_large_scale_fusion_simulation}, just to name a few (see Figure \ref{fig:intro_examples}).
+Other examples include asteroid impact simulations \cite{asteroid_vis_paper}, black hole mergers \cite{blackhole_merger}, and nuclear fusion simulations \cite{in_situ_visualization_large_scale_fusion_simulation}, just to name a few (see Figure \ref{fig:intro_examples}).
 These achievements highlight the transformative power of distributed systems in scientific research.\vskip 6pt
 
 However, programming distributed applications is a challenge of its own, especially for scientists who are not trained in the proper usage of HPC systems.
@@ -58,8 +58,8 @@ Workflows provide a higher-level abstraction that simplifies the process of buil
 By abstracting away much of the low-level complexity, workflows enable domain scientists to focus on the problem at hand rather than the intricacies of parallel programming. 
 This approach significantly reduces the cognitive load required to implement distributed systems and makes parallel computing more accessible to non-computer scientists.\vskip 6pt
 
-In this paper, we introduce \emph{DistributedWorkflows.jl}\cite{DistributedWorkflows}, a Julia package that aims to facilitate the writing of parallel applications. 
-\emph{DistributedWorkflows.jl} is a high-level wrapper around the HPC tool GPI-Space\cite{GSPC}, which allows domain scientists to create efficient parallel applications with a user-friendly interface. 
+In this paper, we introduce \emph{DistributedWorkflows.jl} \cite{DistributedWorkflows}, a Julia package that aims to facilitate the writing of parallel applications. 
+\emph{DistributedWorkflows.jl} is a high-level wrapper around the HPC tool GPI-Space \cite{GSPC}, which allows domain scientists to create efficient parallel applications with a user-friendly interface. 
 This package simplifies the process of building distributed applications by providing a higher-level abstraction, enabling scientists to focus on their research rather than the challenges of parallel programming.\vskip 6pt
 
 The paper is structured as follows:
@@ -74,26 +74,27 @@ Finally, Section \ref{sec:conclusions_and_future_work} briefly describes future 
 \label{sec:related_work}
 
 Julia's rich ecosystem already includes some packages for high-performance computing.
-The most popular of these packages are \emph{Distributed.jl}\cite{Distributedjl} and \emph{Dagger.jl}\cite{dagger1, dagger2}.\vskip 6pt
+The most popular of these packages are \emph{Distributed.jl} \cite{Distributedjl} and \emph{Dagger.jl} \cite{dagger1, dagger2}.\vskip 6pt
 
 \emph{Distributed.jl} offers functionalities for creating and managing multiple Julia processes remotely, enabling distributed and parallel computing. 
[…]
