+++ paper/ref.bib
@@ -51,12 +51,26 @@ author = {},
 }
 
 Citation # 7
-@misc{Introduction_to_petri_nets_book,
-author = {},
-title = {{I}ntroduction to {P}etri {N}ets},
-doi = {10.1007/978-1-4471-4276-8_10},
-howpublished = {\url{https://link.springer.com/chapter/10.1007/978-1-4471-4276-8_10}}
+@Inbook{Introduction_to_petri_nets_book,
+author="Cabasino, M. P.
+and Giua, A.
+and Seatzu, C.",
+editor="Seatzu, C.
+and Silva, M.
+and van Schuppen, J. H.",
+title="Introduction to Petri Nets",
+bookTitle="Control of Discrete-Event Systems: Automata and Petri Net Perspectives",
+year="2013",
+publisher="Springer London",
+address="London",
+pages="191--211",
+abstract="Petri nets are one of the most important discrete event systems formalisms. Three are the main reasons of this. Firstly, they provide a rich family of both logic and timed models, that share a set of formal tools. Secondly, they can be used in all phases of design of a control system. Finally, Petri nets have been successfully used in several research domains, such as max-plus algebra, markovian processes, supervisory control, etc. In this chapter we provide the basic notations and results in this framework, only focusing on a purely logic model called place/transition net.",
+isbn="978-1-4471-4276-8",
+doi="10.1007/978-1-4471-4276-8_10",
+url="https://doi.org/10.1007/978-1-4471-4276-8_10",
+note = {\url{https://doi.org/10.1007/978-1-4471-4276-8_10}}
 }
+
 
 Citation # 8
 @misc{JLD2,
@@ -69,7 +83,7 @@ Citation # 8
 
 Citation # 9
 @misc{Plutojl,
-author = {},
+  author = {},
   title = {Pluto.jl — interactive {J}ulia programming environment},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -84,11 +98,15 @@ Citation # 10
 }
 
 Citation # 11
-@misc{Understanding_Petri_nets_book,
-author = {},
-title = {{U}nderstanding {P}etri {N}ets: {M}odeling {T}echniques, {A}nalysis {M}ethods, {C}ase {S}tudies},
-doi = {10.1007/978-3-642-33278-4},
-howpublished = {\url{https://link.springer.com/book/10.1007/978-3-642-33278-4}}
+@book{Understanding_Petri_nets_book,
+  title = {Understanding Petri Nets: Modeling Techniques,  Analysis Methods,  Case Studies},
+  ISBN = {9783642332784},
+  url = {http://dx.doi.org/10.1007/978-3-642-33278-4},
+  note = {\url{http://dx.doi.org/10.1007/978-3-642-33278-4}},
+  DOI = {10.1007/978-3-642-33278-4},
+  publisher = {Springer Berlin Heidelberg},
+  author = {Reisig,  W.},
+  year = {2013}
 }
 
 Citation # 12
@@ -102,12 +120,13 @@ Citation # 12
 
 Citation # 13
 @inproceedings{gpispace_paper_singular,
-author = {B\"{o}hm, Janko and Fr\"{u}hbis-Kr\"{u}ger, Anne},
+author = {B\"{o}hm, J. and Fr\"{u}hbis-Kr\"{u}ger, A.},
 title = {Massively Parallel Computations in Algebraic Geometry},
 year = {2021},
 isbn = {9781450383820},
 publisher = {Association for Computing Machinery},
 address = {New York, NY, USA},
[…]
