### 2025-11-12 8063c4c
added url for the Los Alamos paper on asteroid impacts
