+++ paper/ref.bib
@@ -294,7 +294,7 @@ url = {http://datascience.dsscale.org/wp-content/uploads/2017/08/VisualizationAn
 year = {2016},
 date = {2016-01-01},
 journal = {2016 ACM/IEEE International Conference for High Performance Computing, Networking, Storage, and Analysis (SC)},
-note = {Winner, Best Scientific Visualization and Data Analytics Showcase; LA-UR-16-26258},
+note = {Winner, Best Scientific Visualization and Data Analytics Showcase; Los Alamos National Laboratory Report LA-UR-16-26258; Copy Available at: \url{https://www.researchgate.net/publication/310747967_Visualization_and_Analysis_of_Threats_from_Asteroid_Ocean_Impacts}},
 keywords = {asteroid, visualization and data analysis},
 pubstate = {published},
 tppubtype = {inproceedings}
