+++ paper/ref.bib
@@ -115,7 +115,8 @@ Citation # 12
   author={Alomairy, R. and Tome, F. and Samaroo, J. and Edelman, A.},
   pages={1--6},
   year={2024},
-  publisher={MIT Open Access Articles}
+  publisher={MIT Open Access Articles},
+  note = {\url{https://www.juliapackages.com/p/dagger}}
 }
 
 Citation # 13
@@ -342,7 +343,8 @@ Citation # 30
   title={{E}fficient {D}ynamic {T}ask {S}cheduling in {H}eterogeneous {E}nvironments with {J}ulia},
   author={Samaroo, J. and Alomairy, R. and Giordano, M. and Edelman, A.},
   year={2024},
-  publisher={MIT Open Access Articles}
+  publisher={MIT Open Access Articles},
+  note = {\url{https://www.juliapackages.com/p/dagger}}
 }
 
 Citation # 31
