### 2026-01-16 #181 @fingolfin
I was informed that there is a new revision of the paper, but the paper attached here seems to be the same that I looked at before. I'd be happy to look at the new version if someone can point me to it.
