### 2026-01-16 FiroozehDastur/DistributedWorkflow.jl#21 @fingolfin
> this package is not intended for production or industry use

Fully understood. However, also academic users must comply with legal aspects and e.g. respect licensing conditions for bundled open source software. I am glad to hear that you now do, especially since you yourself are concerned about "IP misuse and theft".

Regarding the "cautionary “Compatibility Notice” in the README", unfortunately I could not find it. Can you specify in which part exactly it is, or just quote the relevant section?
