### 2026-01-16 #181 @FiroozehDastur
You may be looking at a different branch. In accordance with the JuliaCon proceedings guidelines, all paper-related updates are in the `paper/juliacon2024` branch, under the `paper` folder.
For convenience, the materials are available here:
[paper/juliacon2024 branch](https://github.com/FiroozehDastur/DistributedWorkflows.jl/tree/paper/juliacon2024)

There’s no need to compile the LaTeX sources; an updated PDF is available in the same folder or here:
[paper PDF](https://github.com/FiroozehDastur/DistributedWorkflows.jl/blob/paper/juliacon2024/paper/paper.pdf)
