### 2026-01-19 fad7ba3
Update README.md
Fixed minor typos and outdated commands
