### 2026-01-19 c0a5ac2
Update README.md
Fixed typo
