### 2026-03-06 FiroozehDastur/DistributedWorkflow.jl#21 @fingolfin
Ah I was indeed looking at the `main` branch. Indeed, it has a clear `Compatibility Notice`, thank you!

However, there still seems to be no licensing information regarding the binaries at https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases so I can't tell whether I am e.g. legally allowed to distribute them further.

But inspecting one of these blobs further, it turns out to contain e.g. a binary of GNU m4, but without the required accompanying GPL license. This looks to me like a clear violation of the terms of the GPL -- but perhaps I am missing something? I tried searching this repository for "m4" but the only hit I got was this issue.
