### 2026-03-06 #181 @fingolfin
I've completed my checklist and all seems satisfactory.

(I will mention that I have some concerns with a binary blob that is necessary to run the software and for which licensing terms are unclear, see <https://github.com/FiroozehDastur/DistributedWorkflows.jl/issues/21>, but the license of the Julia code in the repository itself is clearly specified in a license file as "BSD 3-Clause License", so that seems orthogonal to the questions asked by the review checklist)
