### 2026-03-08 #181 @luraess
@FiroozehDastur can you comment on the license topic? Thanks for the revision @fingolfin !
