### 2026-03-08 #181 @mzeyen1985
There are no licensing issues with the binary blob which acts as an installer containing a pre-compiled package repository. According to software licenses only installations need to reproduce the license agreement, while a package repository doesn't qualify as an installation. If any of these third-party installations are in violation of this, I kindly ask @fingolfin to bring this issue up with them.
