### 2026-03-08 FiroozehDastur/DistributedWorkflow.jl#21 @fingolfin
> There are no licensing issues with the binary blob which acts as an installer containing a pre-compiled package repository. According to software licenses only installations need to reproduce the license agreement, while a package repository doesn't qualify as an installation. If any of these third-party installations are in violation of this, I kindly ask @fingolfin to bring this issue up with them.

That's a novel legal theory that I never heard before. Do you have any source to back this up?

My understanding is that the GPL does not care about "installation", only about "distribution". Already its preamble makes this clear in layman's terms:

> For example, if you distribute copies of such a program, whether gratis or for a fee, you must pass on to the recipients the same freedoms that you received. You must make sure that they, too, receive or can get the source code. And you must show them these terms so they know their rights.

Then later in section "6. Conveying Non-Source Forms." it states

> You may convey a covered work in object code form under the terms of sections 4 and 5, provided that you also convey the machine-readable Corresponding Source under the terms of this License, in one of these ways:

... and then it lists 5 ways to comply. Indeed, all major package repositories I know which distribute binaries follow these terms. As far as I can tell, you are not following any of them, but I may be mistaken (your statement however lead me to believe that I am not). BTW, personally I think the list is a bit outdated in how it is phrased, and would consider it sufficient to point to e.g. the GNU file mirrors in case of m4 or gmake or diffutils and other GNU FSF tools.

Now, I think there is no point debating that you *are* distributing copies of the programs, or rather let GitHub distribute them for you; but without attaching any "label" on the outside informing about this. Indeed, I only know that GNU m4 is in there because I downloaded one of the files under <https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/tag/v0.2.1> and figured out how to extract it; there are binaries in there). This is the full list I saw
```
berkeley-db-18.1.40
boost-1.63.0
bzip2-1.0.8
ca-certificates-mozilla-2023-05-30
chrpath-0.16
cmake-3.27.7
curl-8.4.0
diffutils-3.9
distributedworkflow-latest
executor-latest
gdbm-1.23
gmake-4.4.1
gpi-space-23.06
gspcdriver-latest
hwloc-2.9.1
libiconv-1.17
libpciaccess-0.17
libsigsegv-2.14
libssh2-1.11.0
libtool-2.4.7
m4-1.4.19
ncurses-6.4
nghttp2-1.57.0
openssl-3.1.3
perl-5.38.0
pkgconf-1.9.5
readline-8.2
util-macros-1.19.3
xz-5.4.1
zlib-ng-2.1.4
zstd-1.5.5
```

However, I am not sure why we are even arguing, it is so easy for you to comply? Just change the text on <https://github.com/FiroozehDastur/DistributedWorkflows.jl/releases/tag/v0.2.1> to explicitly reproduce the list of software included, and for each link to a place where one can obtain the sources (with exception of your own proprietary material of course, where however you should at least link to suitable licensing terms -- not *source* licensing of course, but one still needs a license to know under which conditions one may use your binaries.)

But from a moral and ethical point of view, I find the situation clear: you should comply and give proper attribution -- just like you rightfully want acknowledgment from people using your IP.
