### 2026-03-08 #181 @fingolfin
First off, I'd like to stress more that I don't even know if any of this is relevant for the *journal submission* under discussion here. I tried to make that clear in my post last night, but want to stress it again: I don't know what the exact rules for these proceedings are; there is a link to JOSS guidelines but my understanding is that only the process part of those are relevant, not necessarily any OSS specific guidance.

That said, I do not agree with @mzeyen1985 assessment of the legal situation at all, and posted an in-depth reply at the original issue <https://github.com/FiroozehDastur/DistributedWorkflows.jl/issues/21>. That seems the most appropriate place. Whether or not it is relevant for the journal submission is up to the editor(s).
