### 2026-04-30 #181 @FiroozehDastur
Hi @luraess, thanks for your message. As my co-author @mzeyen1985 noted, there are no licensing issues on our end (and your bot has also cleared the license). The confusion seems to come from how to query Spack for license information. There are no issues with either the interface or the paper.
