### 2026-04-30 #181 @luraess
@FiroozehDastur before we can finalise and accept the paper, could you address the following remaining bits:
- We need an archive for the paper and repo relating to this publication, possibly on Zenodo (you could link it to the GitHub if handy and make a release to have things synced)
- There is one possibly missing DOI to investigate
- It could be handy to have in the paper somewhere the URL of where to access your software
- You may want to change the rendering of hyperlinks to have them maybe coloured text instead of the ugly boxes
- The bullet points dashes could be changed to shorter ones and detached from the text as well

Once this is fixed, we should be good to go. Thanks!
