+++ paper/ref.bib
@@ -110,14 +110,19 @@ Citation # 11
 }
 
 Citation # 12
-@article{dagger1,
-  title={{D}ynamic {T}ask {S}cheduling with {D}ata {D}ependency {A}wareness using {J}ulia},
+@INPROCEEDINGS{dagger1,
   author={Alomairy, R. and Tome, F. and Samaroo, J. and Edelman, A.},
-  pages={1--6},
+  booktitle={2024 IEEE High Performance Extreme Computing Conference (HPEC)}, 
+  title={Dynamic {T}ask {S}cheduling with {D}ata {D}ependency {A}wareness {U}sing {J}ulia}, 
   year={2024},
-  publisher={MIT Open Access Articles},
-  note = {\url{https://www.juliapackages.com/p/dagger}}
-}
+  volume={},
+  number={},
+  pages={1-7},
+  keywords={Runtime;Processor scheduling;Heuristic algorithms;Linear algebra;Computer architecture;Parallel processing;Dynamic scheduling;Heterogeneous networks;Resource management;High level languages},
+  doi={10.1109/HPEC62836.2024.10938467},
+  note = {\url{https://doi.org/10.1109/HPEC62836.2024.10938467}}
+  }
+
 
 Citation # 13
 @inproceedings{gpispace_paper_singular,
