+++ paper/paper.yml
@@ -16,7 +16,7 @@ authors:
   - name: Mirko Rahn
     affiliation: 2
 affiliations:
-  - name: RPTU Kaiserslautern
+  - name: RPTU Kaiserslautern-Landau
     index: 1
   - name: Fraunhofer ITWM
     index: 2
