+++ paper/header.tex
@@ -5,7 +5,7 @@
 \author[1, 2]{Firoozeh Dastur}
 \author[3]{Max Zeyen}
 \author[2]{Mirko Rahn}
-\affil[1]{RPTU Kaiserslautern-Landau}
+\affil[1]{RPTU Kaiserslautern - Landau}
 \affil[2]{Fraunhofer ITWM}
 \affil[3]{NorthStar Earth \& Space}
 
