+++ paper/paper.tex
@@ -25,10 +25,10 @@
 
 
 \begin{abstract}
-\emph{DistributedWorkflows.jl} is a serializer-independent interface to a distributed task-based workflow management system.
+\texttt{DistributedWorkflows.jl} is a serializer-independent interface to a distributed task-based workflow management system.
 This package aims to simplify the process of writing distributed applications.
 Given a workflow pattern in the form of a Petri net and the Julia code for the workflow tasks, it can be used on a cluster, for example with Slurm, to automate the application's parallel deployment. 
-This makes \emph{DistributedWorkflows.jl} an invaluable addition to Julia's growing ecosystem of high-performance computing.
+This makes \texttt{DistributedWorkflows.jl} an invaluable addition to Julia's growing ecosystem of high-performance computing.
 \end{abstract}
 
 
@@ -65,15 +65,15 @@ Workflows provide a higher-level abstraction that simplifies the process of buil
 By abstracting away much of the low-level complexity, workflows enable domain scientists to focus on the problem at hand rather than the intricacies of parallel programming. 
 This approach significantly reduces the cognitive load required to implement distributed systems and makes parallel computing more accessible to non-computer scientists.\vskip 6pt
 
-In this paper, we introduce \emph{DistributedWorkflows.jl} \cite{DistributedWorkflows}, a Julia package that aims to facilitate the writing of parallel applications. 
-\emph{DistributedWorkflows.jl} is a high-level wrapper around the HPC tool GPI-Space \cite{GSPC}, which allows domain scientists to create efficient parallel applications with a user-friendly interface. 
+In this paper, we introduce \texttt{DistributedWorkflows.jl} \cite{DistributedWorkflows}, a Julia package that aims to facilitate the writing of parallel applications. 
+\texttt{DistributedWorkflows.jl} is a high-level wrapper around the HPC tool GPI-Space \cite{GSPC}, which allows domain scientists to create efficient parallel applications with a user-friendly interface. 
 This package simplifies the process of building distributed applications by providing a higher-level abstraction, enabling scientists to focus on their research rather than the challenges of parallel programming.\vskip 6pt
 
 The paper is structured as follows:
 In Section \ref{sec:related_work}, we provide an overview of related work in distributed computing with Julia and applications that utilize the underlying tool of DistributedWorkflows.
 Section \ref{sec:background} presents background information on Petri nets and GPI-Space. 
 In Section \ref{sec:design_and_method}, we highlight the main features of the package, and describe the design and methodology of the package.
-Section \ref{sec:how_to_use} provides a short description of how to use \emph{DistributedWorkflows.jl}, while Section \ref{sec:limitations} discusses its limitations. 
+Section \ref{sec:how_to_use} provides a short description of how to use \texttt{DistributedWorkflows.jl}, while Section \ref{sec:limitations} discusses its limitations. 
 Finally, Section \ref{sec:conclusions_and_future_work} briefly describes future work and our concluding remarks.
 
 
@@ -81,35 +81,35 @@ Finally, Section \ref{sec:conclusions_and_future_work} briefly describes future 
 \label{sec:related_work}
 
 Julia's rich ecosystem already includes some packages for high-performance computing.
-The most popular of these packages are \emph{Distributed.jl} \cite{Distributedjl} and \emph{Dagger.jl} \cite{dagger1, dagger2}.\vskip 6pt
+The most popular of these packages are \texttt{Distributed.jl} \cite{Distributedjl} and \texttt{Dagger.jl} \cite{dagger1, dagger2}.\vskip 6pt
 
[…]
